requirements.txt
sounddevice
numpy
SpeechRecognition
gTTs
pydub
pyautogui
pytesseract
openai
python-dotenv
elevenlabs
termcolor
ollama
pillow